refactor : step1 페이지 api 활성화
